// src/types/types.ts
export interface CardType {
    id: string;
    title: string;
    description: string;
    priority: string;
}

export interface ColumnType {
    id: string;
    title: string;
    color: string;
    cards: CardType[];
}
// ...
export interface BoardProps {
    columns: ColumnType[];
    onAddCard: (columnId: string, newCard: CardType) => void;
    onDeleteCard: (cardId: string, columnId: string) => void;
    onDeleteColumn: (columnId: string) => void;
    onDragStart: (cardId: string, columnId: string) => void;
    onDrop: (targetColumnId: string, index: number | null) => void;
    dropPosition:{ columnId: string | null;
        index: number | null;
        position: string | null },
    onSetDropPosition:(columnId: string, index: number | null) => void
    onEditCard: (cardId:string, columnId: string, newCard: { title: string; description: string; priority: string }) => void
    onColumnDragStart: (columnId: string) => void
    onColumnDrop: (targetColumnId: string, index: number | null) => void
    onEditColumn: (columnId: string, newColumn: { title: string;}) => void;
}
// ...
export interface DropPositionBoard {
    columnId: string | null;
    index: number | null;
    position: 'before' | 'after' | null;
}